Fix a possible ambiguous call on some platforms

+++ src/lowlevel/Geometry.cpp
@@ -73,7 +73,7 @@ double Geometry::get_distance(int x1, int y1, int x2, int y2) {
 
   int dx = x2 - x1;
   int dy = y2 - y1;
-  return std::sqrt(dx * dx + dy * dy);
+  return std::sqrt((float) (dx * dx + dy * dy));
 }
 
 /**
